refactor for bitmap_pma and pmm

+++ kernel/include/memory/pmm.h
@@ -7,17 +7,23 @@
 #define PMM_FOR_VIRTUAL_MEMORY  1
 #define PMM_FOR_DMA             2
 
-typedef size_t (*PMA_ALLOC_PFN)(void   *a_pma,
-                                size_t  a_size,
-                                size_t *a_physicalAddress);
+typedef uint32 (*PMA_ALLOC_PFN)(
+    void *a_pma,
+    uint64 a_size,
+    uint64 *a_physicalAddress
+);
 
-typedef size_t (*PMA_FREE_PFN)(void  *a_pma,
-                               size_t a_size,
-                               size_t a_physicalAddress);
+typedef uint32 (*PMA_FREE_PFN)(
+    void *a_pma,
+    uint64 a_size,
+    uint64 a_physicalAddress
+);
 
-typedef size_t (*PMA_RESERVE_PFN)(void *a_pma,
-                                  size_t a_size,
-                                  size_t physicalAddress);
+typedef uint32 (*PMA_RESERVE_PFN)(
+    void *a_pma,
+    uint64 a_size,
+    uint64 physicalAddress
+);
 
 struct PMA
 {
@@ -29,15 +35,34 @@ struct PMA
     PMA_RESERVE_PFN reserveFn;
 };
 
-size_t PMM_init(uint8 a_allocatorsNumber);
+uint32 PMM_init(
+    uint8 a_allocatorsNumber
+);
 
-size_t PMM_addAllocator(void *a_allocator, uint8 a_flag,
-                        PMA_ALLOC_PFN a_allocFn, PMA_FREE_PFN a_freeFn, PMA_RESERVE_PFN a_reserveFn);
+uint32 PMM_addAllocator(
+    void *a_allocator,
+    uint8 a_flag,
+    PMA_ALLOC_PFN a_allocFn,
+    PMA_FREE_PFN a_freeFn,
+    PMA_RESERVE_PFN a_reserveFn
+);
 
-size_t PMM_alloc(size_t *a_address, size_t a_size, uint8 a_flag);
+uint32 PMM_alloc(
+    uint64 *a_address,
+    uint64 a_size,
+    uint8 a_flag
+);
 
-size_t PMM_free(size_t a_address, size_t a_size, uint8 a_flag);
+uint32 PMM_free(
+    uint64 a_address,
+    uint64 a_size,
+    uint8 a_flag
+);
 
-size_t PMM_reserve(size_t a_address, size_t a_size, uint8 a_flag);
+uint32 PMM_reserve(
+    uint64 a_address,
+    uint64 a_size,
+    uint8 a_flag
+);
 
 #endif
